requirements.txt
mammoth>=1.6.0
pymupdf4llm>=0.0.17
markdownify>=0.11.6
python-docx>=1.1.0
python-frontmatter>=1.1.0
PyYAML>=6.0.1
anthropic>=0.30.0
requests>=2.31.0
rich>=13.7.0
click>=8.1.7
watchdog>=4.0.0
notebooklm-py>=0.1.0
